requirements.txt
# Core Dependencies
torch>=2.2.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# Federated Learning
opacus>=1.4.0
# pysyft / syft — removed: 0.8.x pins jaxlib>=0.3.14 (unavailable for Python ≥3.11).
# Re-add when a compatible version is available (e.g. syft>=0.9).

# Distributed System & Communication
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.25.0
lz4>=4.3.2
zstandard>=0.22.0
jaeger-client>=4.8.0
py-zipkin>=1.2.0

# Graph Processing
networkx>=3.1
python-igraph>=0.10.6

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.21
redis>=5.0.1

# Monitoring & Metrics
prometheus-client>=0.17.1
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# System metrics for edge nodes
psutil>=5.9.6

# Utilities
loguru>=0.7.2
tqdm>=4.66.1
requests>=2.31.0
aiohttp>=3.8.5

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.17.0

# Container & Deployment
kubernetes>=28.1.0
docker>=6.1.3

# Data Generation
faker>=19.6.2
scipy>=1.11.2
